Feature/refactoring and documentation (#5)
* Split ApiNetworking.swift into dedicated networking and utility files for better modularity

* readme file polish

* Updated README and created a RESOURCES file

---------
