+++ RESOURCES.md
@@ -1,0 +1,82 @@
+//
+//  RESOURCES.md
+//  NasaPodProject
+//
+//  Created by Emmanuel Mathar on 16/01/2025.
+//
+
+# Documentation and Resources Used
+
+This project utilises various resources to achieve its functionality. Below is a list of references used during development, categorised for convenience.
+
+---
+
+## Swift Language and Frameworks
+
+- **Published Property Wrapper**
+  Learn about the `@Published` property wrapper in SwiftUI.
+  [Hacking with Swift](https://www.hackingwithswift.com/quick-start/swiftui/what-is-the-published-property-wrapper)
+
+- **NSError in Swift**
+  Documentation on using `NSError` for error handling.
+  [Apple Developer Documentation](https://developer.apple.com/documentation/foundation/nserror)
+
+- **Enums in Swift**
+  Comprehensive guide on enumerations in Swift.
+  [Swift.org](https://docs.swift.org/swift-book/LanguageGuide/Enumerations.html)
+
+- **Guard Statement**
+  Explore the `guard` statement for improving control flow in Swift.
+  [Swift.org](https://docs.swift.org/swift-book/LanguageGuide/ControlFlow.html)
+
+- **Optional Chaining**
+  Learn about safely unwrapping optionals using optional chaining.
+  [Swift.org](https://docs.swift.org/swift-book/LanguageGuide/OptionalChaining.html)
+
+---
+
+## Networking and API Integration
+
+- **Fetching API Data**
+  A step-by-step guide on fetching data from APIs using `URLSession`.
+  [Medium - Anjali Joshi](https://anjalijoshi2426.medium.com/fetch-data-from-api-using-urlsession-in-swift-3f23f1d35e01)
+  [Luke Roberts Blog](https://lukeroberts.co/blog/fetch-data-api/)
+
+- **URLSession Documentation**
+  Apple's official documentation for the `URLSession` class.
+  [Apple Developer Documentation](https://developer.apple.com/documentation/foundation/urlsession)
+
+- **JSONDecoder Documentation**
+  Decode JSON data seamlessly using `JSONDecoder`.
+  [Apple Developer Documentation](https://developer.apple.com/documentation/foundation/jsondecoder)
+
+- **APOD API GitHub Repository**
+  GitHub repository for NASA's Astronomy Picture of the Day API.
+
+---
+
+## UI and Styling
+
+- **MVVM Architecture**
+  Guide to implementing the MVVM pattern in iOS applications.
+  [Medium - Zeba Yasmeen](https://medium.com/@zebayasmeen76/mvvm-in-ios-swift-6afb150458fd)
+
+- **Styling Text Views in SwiftUI**
+  Tutorial on customising text views with fonts, colours, line spacing, and more.
+  [SwiftUI by Example](https://www.hackingwithswift.com/books/ios-swiftui/resizing-images-to-fit-the-available-space)
+
+- **Screen Size and Image Resizing**
+  Insights on resizing images to fit the available space in SwiftUI.
+  [Hacking with Swift](https://www.hackingwithswift.com/books/ios-swiftui/resizing-images-to-fit-the-available-space)
+
+---
+
+## Video Tutorials and Examples
+
+- **API Calls**
+  A comprehensive YouTube tutorial on making API calls in Swift.
+  [YouTube](https://youtu.be/VWCAa4L-spg?si=hSDcjPWpV2p8yjIr)
+
+---
+
+Feel free to explore these resources for more insights into the project's development and concepts!
